秒杀活动启动修改mongo 修改部分代码规范

+++ framework/src/main/java/cn/lili/cache/impl/RedisCache.java
@@ -207,7 +207,7 @@ public class RedisCache implements Cache {
         RedisAtomicLong entityIdCounter = new RedisAtomicLong(key, redisTemplate.getConnectionFactory());
         Long increment = entityIdCounter.getAndIncrement();
         //初始设置过期时间
-        if ((null == increment || increment.longValue() == 0) && liveTime > 0) {
+        if ((null == increment || increment == 0) && liveTime > 0) {
             entityIdCounter.expire(liveTime, TimeUnit.SECONDS);
         }
